refactor: move publishing orchestration tests
